ipacm: Added changes to check sw allow rule or not.
Added changes to check rule is sw allow rule or not at the time
of adding to the temp cache.

(cherry picked from commit 7412b71686039c2bc4f3adffa02d3d50fe4dc676)

+++ ipacm/src/IPACM_Conntrack_NATApp.cpp
@@ -320,24 +320,30 @@ int NatApp::AddPdn(uint32_t pub_ip, uint8_t mux_id, bool is_sta)
 			nat_rule.s = cache[cnt].s;
 			nat_rule.pdn_index = pdn_index;
 			cache[cnt].pdn_index = pdn_index;
-
-			if(ipa_nat_add_ipv4_rule(nat_table_hdl, &nat_rule, &cache[cnt].rule_hdl) < 0)
+			if(cache[cnt].sw_allow)
 			{
-				IPACMERR("unable to add the rule delete from cache\n");
-				memset(&cache[cnt], 0, sizeof(cache[cnt]));
-				curCnt--;
-				continue;
+				IPACMDBG("SW allow entry so not adding to HW\n");
 			}
-			IPACMDBG("cache entry %d rule handle %d\n", cnt, cache[cnt].rule_hdl);
-			cache[cnt].enabled = true;
+			else
+			{
+				if(ipa_nat_add_ipv4_rule(nat_table_hdl, &nat_rule, &cache[cnt].rule_hdl) < 0)
+				{
+					IPACMERR("unable to add the rule delete from cache\n");
+					memset(&cache[cnt], 0, sizeof(cache[cnt]));
+					curCnt--;
+					continue;
+				}
+				IPACMDBG("cache entry %d rule handle %d\n", cnt, cache[cnt].rule_hdl);
+				cache[cnt].enabled = true;
 
-			IPACMDBG("new pdn added below rule successfully\n");
-			iptodot("Private IP", nat_rule.private_ip);
-			iptodot("Target IP", nat_rule.target_ip);
-			IPACMDBG("Private Port:%d \t Target Port: %d\n", nat_rule.private_port, nat_rule.target_port);
-			IPACMDBG("Public Port:%d\n", nat_rule.public_port);
-			IPACMDBG("protocol: %d\n", nat_rule.protocol);
-			IPACMDBG("pdn index: %d\n", nat_rule.pdn_index);
+				IPACMDBG("new pdn added below rule successfully\n");
+				iptodot("Private IP", nat_rule.private_ip);
+				iptodot("Target IP", nat_rule.target_ip);
+				IPACMDBG("Private Port:%d \t Target Port: %d\n", nat_rule.private_port, nat_rule.target_port);
+				IPACMDBG("Public Port:%d\n", nat_rule.public_port);
+				IPACMDBG("protocol: %d\n", nat_rule.protocol);
+				IPACMDBG("pdn index: %d\n", nat_rule.pdn_index);
+			}
 		}
 
 	}
@@ -1649,6 +1655,10 @@ void NatApp::CacheEntry(const nat_table_entry *rule)
 			cache[cnt].uc_activation_index = rule->uc_activation_index;
 			cache[cnt].ucp = rule->ucp;
 			cache[cnt].s = rule->s;
+			if(ChkSWAllow(rule))
+			{
+				cache[cnt].sw_allow = true;
+			}
 #ifdef FEATURE_VLAN_MPDN
 			cache[cnt].pdn_index = rule->pdn_index;
 #endif
